fix(jwttoken): correct JwtTokenEntity id type to match database schema
- Change id field type from Long to Integer
- Resolves schema validation error: found [serial (Types#INTEGER)], expecting [bigint (Types#BIGINT)]
- Aligns JPA entity definition with existing database column type (serial)

+++ src/main/java/com/spacecodee/securityspacee/jwttoken/infrastructure/persistence/jpa/JwtTokenEntity.java
@@ -2,8 +2,6 @@ package com.spacecodee.securityspacee.jwttoken.infrastructure.persistence.jpa;
 
 import java.time.Instant;
 
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
 import org.jspecify.annotations.NonNull;
 import org.jspecify.annotations.Nullable;
 
@@ -18,8 +16,10 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Table(name = "jwt_token")
@@ -31,7 +31,7 @@ public class JwtTokenEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "jti", nullable = false, unique = true)
     private @NonNull String jti;
